--- ct-app/postman/task_status.py
@@ -1,15 +1,0 @@
-from enum import Enum
-
-
-class TaskStatus(Enum):
-    """
-    Enum to represent the status of a task. This status is also used when creating a
-    task in the outgoing queue.
-    """
-
-    DEFAULT = "DEFAULT"
-    SUCCESS = "SUCCESS"
-    RETRIED = "RETRIED"
-    SPLIT = "SPLIT"
-    TIMEOUT = "TIMEOUT"
-    FAILED = "FAILED"
